requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-appconf==1.0.5
django-autoslug==1.9.8
django-filebrowser==4.0.2
django-grappelli==3.0.4
django-imagekit==4.1.0
django-tinymce==3.5.0
pilkit==2.0
Pillow==9.4.0
six==1.16.0
sqlparse==0.4.3
